Add test for can fetch one product and passes

+++ database/migrations/2018_09_22_184106_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->increments('id');
             $table->string('name')->index();
             $table->longText('description');
-            $table->integer('price');
+            $table->float('price');
             $table->integer('quantity');
             $table->timestamps();
             $table->softDeletes();
